feat: wrapup haystack service; update test structure
