--- src/deepset_mcp/api/haystack_service/models.py
@@ -1,11 +1,0 @@
-from typing import Any
-
-from pydantic import BaseModel
-
-
-class HaystackComponentSchema(BaseModel):
-    """Model for the Haystack component schema response."""
-
-    model_config = {"extra": "allow"}
-
-    __root__: dict[str, Any]
